main updated

+++ src/main/java/com/solvd/navigatorProject/dao/interfaces/IVertexDao.java
@@ -4,6 +4,6 @@ import com.solvd.navigatorProject.binary.Vertex;
 
 public interface IVertexDao extends IEntityDao<Vertex>{
 
-    double getDistanceById(long id);
+    long getDistanceById(long id);
 
 }
